requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
